requirements.txt
Django~=4.2.11
Pillow==10.3.0
django-admin-sortable2==2.2.3